// src/app/models/products.ts
// export interface Product {
//   id: number;
//   title: string;
//   price: number;
//   description: string;
//   images: string[];
  // creationAt: Date;
  // updatedAt: Date;
  // category: Category;
// }
// export interface Category {
//   id: number;
//   name: string;
//   image: string;
//   creationAt: Date;
//   updatedAt: Date;
// }
export interface Product {
  id: number;
  title: string;
  price: number;
  description: string;
  // category: string;
  image: string;
  // rating: Rating;
}

// export interface Rating {
//   rate: number;
//   count: number;
// }